Add support for storing the SMP Long Term Key
This adds support for receiving keys with arbitrary data following
it, this is because for SMP we may have different types of keys that
have different kind of data associated with them.

And we keep that associated data opaque, as only the communication
functions need to be aware of its contents, for storage there's no
need to export that information.