observer: Add watchers and filters for observers
The adapter will keep a list of observers to listening for advertising.
The new observer is added on the list when it is registered.

When an observer is registered, the filter must be stored to be used
after receiving a new advertising. The observer only will received the
data from advertising if the filter matches.