core: add compare function for bdaddr in a struct btd_device
This is a utility function similar to device_address_cmp but comparing
bdaddr instead of the string representing the address. This way is
possible to avoid allocating two buffers to temporarily hold the strings,
two sprintf() calls to generate the strings from the bdaddr arrays, and
a string comparison, substituting all of it for one memcmp() call.