hog: Adds initial files for suspend support
This patch adds the initial files to support HoG suspend. The first
supported backend is UPower. The suspend concept for HoG is
implementation specific. For UPower, Suspend/Resume will be triggered
by "Sleeping" and "Resuming" signals.

When setting the Control Point, the report device can execute actions to
save power. eg: Reduce the cycle of the key press detection or disable
LEDs.