FIXUP: Keep const data pointer for hooks for now
While we do not need hooks that modify HCI packets, let's keep the
pointers const to avoid ugly casts.

Once we require hooks that modify HCI packets, we need to change lower
layers to pass non-const pointers to avoid cast.

NOTE: this fixup should be split to apply to commits that it modifies.