gatt: Add Write Request
This patch adds the support for ATT Write Request operation in the
attribute server. The attribute write callback calls the DBusProxy
write callback informing the result callback for synchronous operation
result. From the Proxy perspective, the Write Request is handled in
similar manner to Write Command, except for the result callback that
needs to called to report the operation status for Write Request
operations.