gatt-dbus: Implement read descriptor callback
This patch implements the descriptor read callback for characteristics
provided by an external implementation. When called, it gets the "Value"
attribute of the descriptor proxy object. Proxy objects like any other
DBus object reports changes through DBus Properties. If "Value" property
is not available means that the descriptor is not readable.