gatt: Add mandatory GAP Appearance characteristic
There is no direct mapping between BR/EDR Class of Device (CoD) and GAP
Appearance, therefore a simple mapping is implemented that recognizes
whether the system is a computer or a phone.

In future, a better approach could be to map GATT service UUIDs
registered on the local database to more specific appearance values, or
allow this field to be freely defined over D-Bus.