gatt: Add helper for writing characteristics
This patch adds the btd_gatt_write_attribute() helper. At the moment, it
is restricted to local attributes only. It is an abstraction to allow
external servers implementations (DBus Proxies) sending notification or
indication. When PropertiesChanged is received, connected peers with CCC
enabled will get notification or indication. The ATT transaction is
hidden, Proxies don't get confirmation when and who received the values.